CustomMessageBox: abstract

+++ Controls/MavlinkNumericUpDown.cs
@@ -137,7 +137,7 @@ namespace MissionPlanner.Controls
             {
                 if (
                     CustomMessageBox.Show(ParamName + " Value out of range\nDo you want to accept the new value?",
-                        "Out of range", MessageBoxButtons.YesNo) == DialogResult.Yes)
+                        "Out of range", MessageBoxButtons.YesNo) == (int)DialogResult.Yes)
                 {
                     base.Maximum = decimal.Parse(value);
                     base.Value = decimal.Parse(value);
